requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
faker==20.1.0
pytest-repeat==0.9.1
pytest-xdist==3.5.0
flake8==6.1.0
black==23.11.0
pytest-timeout==2.2.0